+++ messages/index.js
@@ -1,4 +1,11 @@
 const createInvoiceMessage = require('./create_invoice_message');
+const createTradeMessage = require('./create_trade_message');
 const failureMessage = require('./failure_message');
+const settleTradeMessage = require('./settle_trade_message');
 
-module.exports = {createInvoiceMessage, failureMessage};
+module.exports = {
+  createInvoiceMessage,
+  createTradeMessage,
+  failureMessage,
+  settleTradeMessage,
+};