+++ post/index.js
@@ -2,9 +2,11 @@ const notifyOfForwards = require('./notify_of_forwards');
 const postChainTransaction = require('./post_chain_transaction');
 const postClosedMessage = require('./post_closed_message');
 const postCreatedInvoice = require('./post_created_invoice');
+const postCreatedTrade = require('./post_created_trade');
 const postOpenMessage = require('./post_open_message');
 const postSettledInvoice = require('./post_settled_invoice');
 const postSettledPayment = require('./post_settled_payment');
+const postSettledTrade = require('./post_settled_trade');
 const postUpdatedBackup = require('./post_updated_backup');
 const sendMessage = require('./send_message');
 
@@ -13,9 +15,11 @@ module.exports = {
   postChainTransaction,
   postClosedMessage,
   postCreatedInvoice,
+  postCreatedTrade,
   postOpenMessage,
   postSettledInvoice,
   postSettledPayment,
+  postSettledTrade,
   postUpdatedBackup,
   sendMessage,
 };